create date block widget
